requirements.txt
flask==2.3.3
openai==1.3.5
python-dotenv==1.0.0
requests==2.31.0
pandas==2.1.3
numpy==1.24.3
jinja2==3.1.2
gunicorn==21.2.0
plotly==5.17.0
streamlit==1.28.1
